+++ clients/privacy-center/pages/index.tsx
@@ -1,4 +1,11 @@
-import { Flex, Heading, Text, Stack, useToast } from "@fidesui/react";
+import {
+  Flex,
+  Heading,
+  Text,
+  Stack,
+  useToast,
+  useDisclosure,
+} from "@fidesui/react";
 import React, { useEffect, useState } from "react";
 import type { NextPage } from "next";
 import { useRouter } from "next/router";
@@ -16,6 +23,11 @@ import { useGetIdVerificationConfigQuery } from "~/features/id-verification";
 import PrivacyCard from "~/components/PrivacyCard";
 import ConsentCard from "~/components/consent/ConsentCard";
 import { useConfig } from "~/features/common/config.slice";
+import { useSubscribeToPrivacyExperienceQuery } from "~/features/consent/hooks";
+import { useAppSelector } from "~/app/hooks";
+import { selectPrivacyExperience } from "~/features/consent/consent.slice";
+import NoticeEmptyStateModal from "~/components/modals/NoticeEmptyStateModal";
+import { selectIsNoticeDriven } from "~/features/common/settings.slice";
 
 const Home: NextPage = () => {
   const router = useRouter();
@@ -47,6 +59,24 @@ const Home: NextPage = () => {
   } = useConsentRequestModal();
   let isConsentModalOpen = isConsentModalOpenConst;
   const getIdVerificationConfigQuery = useGetIdVerificationConfigQuery();
+
+  // Subscribe to experiences just to see if there are any notices.
+  // The subscription automatically handles skipping if overlay is not enabled
+  useSubscribeToPrivacyExperienceQuery();
+  const noticeEmptyStateModal = useDisclosure();
+  const experience = useAppSelector(selectPrivacyExperience);
+  const isNoticeDriven = useAppSelector(selectIsNoticeDriven);
+  const emptyNotices =
+    experience?.privacy_notices == null ||
+    experience.privacy_notices.length === 0;
+
+  const handleConsentCardOpen = () => {
+    if (isNoticeDriven && emptyNotices) {
+      noticeEmptyStateModal.onOpen();
+    } else {
+      onConsentModalOpen();
+    }
+  };
 
   useEffect(() => {
     if (getIdVerificationConfigQuery.isError) {
@@ -88,7 +118,7 @@ const Home: NextPage = () => {
         title={config.consent.button.title}
         iconPath={config.consent.button.icon_path}
         description={config.consent.button.description}
-        onOpen={onConsentModalOpen}
+        onOpen={handleConsentCardOpen}
       />
     );
     if (router.query?.showConsentModal === "true") {
@@ -177,6 +207,11 @@ const Home: NextPage = () => {
         setConsentRequestId={setConsentRequestId}
         isVerificationRequired={isVerificationRequired}
         successHandler={consentModalSuccessHandler}
+      />
+
+      <NoticeEmptyStateModal
+        isOpen={noticeEmptyStateModal.isOpen}
+        onClose={noticeEmptyStateModal.onClose}
       />
     </main>
   );